+++ source/io/UdpClient.h
@@ -20,7 +20,7 @@ public:
 
     // TODO: implement???
     //void send_data(const DataChunk& data_chunk, std::uint32_t host, std::uint16_t port, EndSendCallback callback = nullptr);
-    IO_DLL_PUBLIC void send_data(std::shared_ptr<const char> buffer, std::size_t size, std::uint32_t host, std::uint16_t port, EndSendCallback callback = nullptr);
+    IO_DLL_PUBLIC void send_data(std::shared_ptr<const char> buffer, std::uint32_t size, std::uint32_t host, std::uint16_t port, EndSendCallback callback = nullptr);
     IO_DLL_PUBLIC void send_data(const std::string& message, std::uint32_t host, std::uint16_t port, EndSendCallback callback = nullptr);
 
     IO_DLL_PUBLIC void schedule_removal() override;
